third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-transform-function-expected.txt

This is a testharness.js-based test.
FAIL Animating a custom property of type <transform-function> assert_equals: expected "translateX(150px)" but got "translateX(200px)"
FAIL Animating a custom property of type <transform-function> with a single keyframe assert_equals: expected "translateX(150px)" but got "translateX(200px)"
FAIL Animating a custom property of type <transform-function> with additivity assert_equals: expected "translateX(350px)" but got "translateX(300px)"
FAIL Animating a custom property of type <transform-function> with a single keyframe and additivity assert_equals: expected "translateX(250px)" but got "translateX(300px)"
FAIL Animating a custom property of type <transform-function> with iterationComposite assert_equals: expected "translateX(250px)" but got "translateX(100px)"
Harness: the test ran to completion.
